Fix estimation with individual weights
Necessary to update nl and individual weights vector to double type

+++ R/itemfit.R
@@ -116,7 +116,7 @@ NULL
 #'@export
 itemfit = function(data, dim, model = "2PL", EMepsilon = 1e-4, clusters = NULL,
 				  quad_tech = NULL, quad_points = NULL, 
-				  individual_weights = as.integer(c()),
+				  individual_weights = as.numeric(c()),
 				  initial_values = NULL, SD = FALSE,
 				  verbose = TRUE, save_time = TRUE ) {
