+++ R/itemfitbayesian.R
@@ -35,7 +35,7 @@
 #'@export
 itemfit.bayesian = function(data, dim, model = "2PL", EMepsilon = 1e-4, clusters = NULL,
 				  quad_tech = NULL, quad_points = NULL, 
-				  individual_weights = as.integer(c()),
+				  individual_weights = as.numeric(c()),
 				  initial_values = NULL, noguessing = TRUE, 
 				  verbose = TRUE, save_time = TRUE ) {
