+++ src/polytomous/type/estimationdata.h
@@ -28,7 +28,7 @@ public:
 	matrix<char> *dataset; /**< Matrix of answers*/
 	int d; /**< Dimension*/
 	matrix<char> Y; /**< Matrix of response patterns*/
-	std::vector<int> nl; /**< Frequencies of each response pattern*/
+	std::vector<double> nl; /**< Frequencies of each response pattern*/
 	int N; /**< Number of examines*/
 	int s; /**< Number of response patterns*/
 	int p; /**< Number of items*/
